refactor(ARCHITECT-002): split AdminCustomers — extract 3 modals + constants
- Extract CustomerModal (~352 lines) to components/customers/CustomerModal.jsx
- Extract CustomerDetailsModal + PortalSettingsTab (~380 lines) to components/customers/CustomerDetailsModal.jsx
- Extract ImportCSVModal (~344 lines) to components/customers/ImportCSVModal.jsx
- Extract STATUS_OPTIONS, getStatusStyle, formatPhoneNumber to constants.js
- AdminCustomers.jsx: 1511 → 404 lines
